fixed the tx id in the replication api
The localTxId was showing the next tx id which was confusing because it did not match with the id reported by the slave when the master and slave were in sync.
Now they both show the same tx id which is the id of the last transaction.

Note: the nextTxId which is the id of the next transaction expected by the slave stays unchanged.