CCSDS TM index rebuild fix
The rebuild was not starting from the good position (was assuming that the index records are the first one in the database). The result is that the rebuild may have removed the first header record or not remove all the records.
In the new version when a rebuild of the whole index is requested, remove the whole index instead of removing records apid by apid. This should correct errors caused by the previous rebuild.

Also removed some unused methods.