Introducing the RocksDB merge operator for the parameter archive
The operator is able to merge segments into intervals when compacting the database. It also merges them on the fly when retrieving data before the database is compacted. This allows to use less memory in Yamcs during the filling process.

Taking advantage of the breaking change, fix also the long standing issue of non-negative tiemstamps not sorting properly in the parameter archive