dynamic command release
The processor used to release commands in a set of streams configured in the instance.yaml streamConfig section using a list of patterns for each stream.
This change allows to add at runtime additional streams with custom command matchers. It is used by the Yamcs Gateway link to register direct commnads that are forwarded to the YGW nodes.