fix for multiple values with the same timestamp
There may be cases when multiple values with the same timestamp for one parameter are generated. For example when using repeated structures n*param1 in the packet definition.

The ParameterArchive was in principle supporting this case after the introduction of the gaps (the subsequent values are just stored in the same segment with the same timestamp and in position of all the other parameters in the group there is a gap). However during buildup it wasn't working correctly. This has been fixed.

In addition the ArrayParameterCache also did not return multiple values although they were stored. This is fixed as well.