Advance event level migration
For a long time we've had two conflicting
event levels:

* INFO, WARNING, ERROR
* INFO, WATCH, WARNING, DISTRESS, CRITICAL, SEVERE.

This commit advances the deprecation away from
"INFO, WARNING, ERROR".

In a future commit, probably we'll need to split EventSeverity
enum, between the DB and the API, to apply a correct
conversion of old events.

The current commit serves only to add a new state
with serialization id 4 (temporarily named "WARNING_NEW"),
so that Protobuf clients can start to make
preparations for a future migration step.