Bug fix parameter archive rebuilding
When rebuilding the parameter archive the gaps segment was not properly deleted because this check

  if (rdb.get(gapKey) != null) {
      writeBatch.delete(cfh, gapKey);
  }

should have been:

  if (rdb.get(cfh, gapKey) != null) {
      writeBatch.delete(cfh, gapKey);
  }

This caused in certain circumstances the gap segments to be appended to eachohther.
This commit fixes that  bug.

An additional bug was found which caused a rebuild to not remove a (orphan) parameter which was written in a previous build of the parameter archive but disappeared in the meanwhile from the stream archive.

In addition:
 - added a consistency check that the gaps size + values size = timestamps size when reading the archive segments.
 - added some logging to allow easier debugging in the future
 - fixed unit test and added a test for the orphan parameter