Exclude maxGap length from pre-gap range
If a gap is injected due to the maxGap being
exceeded, the maxGap time should not be
included in the prior range.

At the same time remove also the expiration-
related duration from the trailing gap.

-if- we were to include it, it should be
the actual rateInStream. Not including the
multiplier. Currently such information is
not available at pval level.

This commit also increases the allowed
maxRanges property from 500 to 5000.