fix replay when the tm table is not available
It is possible to run Yamcs without receiving any tm packets. In that case there would be no tm table and the replays were failing because of that.

This fixes also the case when there is a tm table but no pp. Same for cmdhistory although not sure command history replay is really used.