calibrator improvements
Do not restrict calibrators to numeric values as per XTCE. Instead allow calibrating any raw value type to any engineering value type.
Renamed ProcessingData to ProcessingContext because the name "ProcessingData" is too similar to "ProcessorData".
Added a generationTime in the ProcessingContext to be used by the time calibrator algorithms that need to correlate on-board times to UTC.