added support for member aggregate references
When sending a command it is allowed now to use arguments of type aggregates whose members are arrays whose length is specified also as part of the aggregates.

For example:

TC1(arg1: {
   n: 3,
   a: [1,2,3]
})

The length of a is specified as referencing n.
Yamcs will also allow to send the command without specifying any value for n, it will be deduced from the length of the array a.