fix race conditions in replication
The replication master was not properly keeping track on which replication files are in use causing crashes if a file that was being used to send data to a slave was closed.
Introduced a reference counted version of ManagedReplicationFile to solve this problem. Also moved the locking logic from the ReplicationFile to this new class to simplify the ReplicationFile class