TM packet metadata
Allow packet pre-processors and other components to add metadata to telemetry packets. The metadata is stored as parameters in the tm table and also alongside (in the same group) the other parameters extracted from the archive in the parameter archive.

Use the new functionality to save the id of the VC (Virtual Channel) the packet has been extracted from, in the CCSDS frame processors